Refactor the terminal view specific for email; Refactor the folder structure to aggregate factor related view into its own folder.
