+++ src/v2/view-builder/views/email/TerminalTransferedEmailView.js
@@ -1,0 +1,11 @@
+import TerminalView from '../TerminalView';
+
+const Body = TerminalView.prototype.Body.extend({
+  title () {
+    return 'You can close this window';
+  },
+});
+
+export default TerminalView.extend({
+  Body,
+});
